% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epistemic-package.R
\docType{package}
\name{epistemic-package}
\alias{epistemic}
\alias{epistemic-package}
\title{epistemic: Tidy Population Modeling of Epidemics}
\description{
The epistemic package is designed to bring population-level epidemic modeling
  into the tidyverse. It is based on the epidemia package and relies on this
  package for most model implementation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jesse-smith/epistemic}
  \item Report bugs at \url{https://github.com/jesse-smith/epistemic/issues}
}

}
\author{
\strong{Maintainer}: Jesse Smith \email{jessesmith96.js@gmail.com} (\href{https://orcid.org/0000-0001-7375-2463}{ORCID})

}
\keyword{internal}
